clearAnnotation & setCenter in AnnotationView
Added a method to clear the annotation context from the annotation view for the purpose of reuse.

The setCenter method is used to figure out where the view is on the screen and if it should show. It is used in the update method by the manager when looping throught the annotations.