MapViewManager Update
Finished the enqueue and dequeue methods. the update method will ask the user for a view using the viewForAnnotation method. It will check if the view is part of the visible area of the map.

Next time we will add the event piping from the map to the touchable wrapper.

Max says we are done in a weird kinda way.

WE SHALL SEE!!! :P